+++ map.h
@@ -19,33 +19,33 @@ public:
   Map(std::ifstream& infile, int resolution);	//constructor, takes in binary map file .pgm
  // ~Map();					//destructor
   Map();
-  void setGridValue(int value, int i, int j);	//setter for value in position ij of the grid
-  int getGridValue(int i, int j) const;		//getter for value in position ij of the grid
-  int getGridValue(int i) const;			//getter for value in position i of the grid vector
-  int getMapValue(int i, int j);		//getter for value in position ij of the map
-  int getNumGridRows() const;
-  int getNumGridCols() const;
-  int getNumRows();
-  int getNumCols();
+  void setGridValue(int value, long i, long j);	//setter for value in position ij of the grid
+  int getGridValue(long i, long j) const;		//getter for value in position ij of the grid
+  int getGridValue(long i) const;			//getter for value in position i of the grid vector
+  int getMapValue(long i, long j);		//getter for value in position ij of the map
+  long getNumGridRows() const;
+  long getNumGridCols() const;
+  long getNumRows();
+  long getNumCols();
   void findFreeEdges(int cX, int cY);
   void addEdgePoint(int x, int y);
   //std::vector<int> getMap();
   //std::vector<int> getGrid();
-  std::vector<vector<int> > getMap2D();
-  std::vector<int> grid;			//vector containing the map as grid of cells sized 1 square metre
-  int numGridRows;
-  int numGridCols;
+  std::vector<vector<long> > getMap2D();
+  std::vector<long> grid;			//vector containing the map as grid of cells sized 1 square metre
+  long numGridRows;
+  long numGridCols;
   Pose getRobotPosition();
   long getTotalFreeCells();
   
   
 protected:
-  std::vector<int> map;				//vector containing the original map as binary matrix (0 -> free, 1 -> obstacle)
+  std::vector<long> map;				//vector containing the original map as binary matrix (0 -> free, 1 -> obstacle)
   void createMap(std::ifstream& infile);
   void createGrid(int resolution);
   void createNewMap();
-  int numRows;
-  int numCols;
+  long numRows;
+  long numCols;
   std::vector<std::pair<int, int> > edgePoints;
   long totalFreeCells;
   void decreaseFreeCells();
