+++ newray.cpp
@@ -14,12 +14,12 @@ NewRay::NewRay()
 
 
 //Check if a cell is candidate position: return 1 if the cell is adjacent to at least one free cell, 0 otherwise
-int NewRay::isCandidate(const Map &map, int i, int j)
+int NewRay::isCandidate(const Map &map, long i,long  j)
 {
   int candidate = 0;
-  int r = i;
-  int s = j;
-  int minR = r - 1, maxR = r + 1, minS = s -1, maxS = s + 1;
+  long r = i;
+  long s = j;
+  long minR = r - 1, maxR = r + 1, minS = s -1, maxS = s + 1;
   if(minR < 0) minR = 0;
   if(minS < 0) minS = 0;
   if(maxR > map.getNumGridRows()) maxR = map.getNumGridRows();
@@ -39,7 +39,7 @@ int NewRay::isCandidate(const Map &map, int i, int j)
 
 
 //finds the candidate positions: cells already scanned in range of the robot which are adjacent to at least one free cell
-void NewRay::findCandidatePositions(Map &map, int posX, int posY, int orientation, double FOV, int range)
+void NewRay::findCandidatePositions(Map &map, long posX, long posY, int orientation, double FOV, int range)
 {
   NewRay::numGridRows = map.getNumGridRows();
   
@@ -59,10 +59,10 @@ void NewRay::findCandidatePositions(Map &map, int posX, int posY, int orientatio
   //std::cout << std::endl << "StartingPhi: " << startingPhi << " EndingPhi: " << endingPhi <<std::endl;
     
   //select the portion of map to be scanned
-  int minI = posX - range;
-  int maxI = posX + range;
-  int minJ = posY - range;
-  int maxJ = posY + range;
+  long minI = posX - range;
+  long maxI = posX + range;
+  long minJ = posY - range;
+  long maxJ = posY + range;
   
   if(minI < 0) minI = 0;
   if(minJ < 0) minJ = 0;
@@ -70,9 +70,9 @@ void NewRay::findCandidatePositions(Map &map, int posX, int posY, int orientatio
   if(maxJ > map.getNumGridCols()) maxJ = map.getNumGridCols();
   
   //scan the cells in the selected portion of the map
-  for(int i = minI; i <= maxI; ++i)
+  for(long i = minI; i <= maxI; ++i)
   {
-    for(int j = minJ; j <=maxJ; ++j)
+    for(long j = minJ; j <=maxJ; ++j)
     {
       
       double distance = sqrt((i - posX)*(i - posX) + (j - posY)*(j - posY));
@@ -118,16 +118,16 @@ void NewRay::findCandidatePositions(Map &map, int posX, int posY, int orientatio
 	    //not needed, but left anyway
 	    if(curX < 0 || curX > map.getNumGridRows() || curY < 0 || curY > map.getNumGridCols()) hit = 1;        
       
-	    if(map.getGridValue((int)curX, (int)curY) == 1) 
+	    if(map.getGridValue((long)curX, (long)curY) == 1) 
 	    {
 	      hit = 1;		//hit set to 1 if an obstacle is found
 	      //std::cout << "HIT! cell: " << j << " " << i << " Hit point: " << curY << " " << curX << std::endl;
 	    }
 	    
 	    
-	    if((int)curX == i && (int)curY == j)	//if the free cell is reached, save it as edge point and stop the ray.
+	    if((long)curX == i && (long)curY == j)	//if the free cell is reached, save it as edge point and stop the ray.
 	    {
-	      std::pair<int, int> temp = std::make_pair(i, j);
+	      std::pair<long,long> temp = std::make_pair(i, j);
 	      NewRay::edgePoints.push_back(temp);
 	      //std::cout << "Cell scanned: " << (int)curY << " " << (int)curX << std::endl;
 	      hit = 1;
@@ -142,7 +142,7 @@ void NewRay::findCandidatePositions(Map &map, int posX, int posY, int orientatio
 }
 
 
-vector< std::pair<int, int> > NewRay::getCandidatePositions()
+vector< std::pair<long,long> > NewRay::getCandidatePositions()
 {
   return NewRay::edgePoints;
 }
@@ -156,7 +156,7 @@ void NewRay::emptyCandidatePositions()
 //calculate the sensing time of a possible scanning operation, returns the minimum FOV required to scan all the free cells from the considered pose
 //ATTENTION: the FOV is always centered in the orientation of the robot
 //ATTENTION: in order to optimize the computing time, this method should be fused with the information gain one
-double NewRay::getSensingTime(const Map &map, int posX, int posY, int orientation, double FOV, int range)
+double NewRay::getSensingTime(const Map &map, long posX,long posY, int orientation, double FOV, int range)
 {
   NewRay::numGridRows = map.getNumGridRows();
   
@@ -179,10 +179,10 @@ double NewRay::getSensingTime(const Map &map, int posX, int posY, int orientatio
   //std::cout << std::endl << "StartingPhi: " << startingPhi << " EndingPhi: " << endingPhi <<std::endl;
     
   //select the portion of map to be scanned
-  int minI = posX - range;
-  int maxI = posX + range;
-  int minJ = posY - range;
-  int maxJ = posY + range;
+  long minI = posX - range;
+  long maxI = posX + range;
+  long minJ = posY - range;
+  long maxJ = posY + range;
   
   if(minI < 0) minI = 0;
   if(minJ < 0) minJ = 0;
@@ -190,9 +190,9 @@ double NewRay::getSensingTime(const Map &map, int posX, int posY, int orientatio
   if(maxJ > map.getNumGridCols()) maxJ = map.getNumGridCols();
   
   //scan the cells in the selected portion of the map
-  for(int i = minI; i <= maxI; ++i)
+  for(long i = minI; i <= maxI; ++i)
   {
-    for(int j = minJ; j <=maxJ; ++j)
+    for(long j = minJ; j <=maxJ; ++j)
     {
       
       double distance = sqrt((i - posX)*(i - posX) + (j - posY)*(j - posY));
@@ -231,13 +231,13 @@ double NewRay::getSensingTime(const Map &map, int posX, int posY, int orientatio
 	    //not needed, but left anyway
 	    if(curX < 0 || curX > map.getNumGridRows() || curY < 0 || curY > map.getNumGridCols()) hit = 1;        
       
-	    if(map.getGridValue((int)curX, (int)curY) == 1) 
+	    if(map.getGridValue((long)curX, (long)curY) == 1) 
 	    {
 	      hit = 1;		//hit set to 1 if an obstacle is found
 	      //std::cout << "HIT! cell: " << j << " " << i << " Hit point: " << curY << " " << curX << std::endl;
 	    }
 
-	    if((int)curX == i && (int)curY == j)	//free cell reached, check if the slope is a candidate for first or last ray
+	    if((long)curX == i && (long)curY == j)	//free cell reached, check if the slope is a candidate for first or last ray
 	    {
 	      if(phiFound == 0)		//enters if it is the first free cell found
 	      {
@@ -275,7 +275,7 @@ double NewRay::getSensingTime(const Map &map, int posX, int posY, int orientatio
 
 
 //perform the sensing operation by setting the value of the free cell scanned to 2
-void NewRay::performSensingOperation(dummy::Map &map, int posX, int posY, int orientation, double FOV, int range)
+void NewRay::performSensingOperation(dummy::Map &map, long posX, long posY, int orientation, double FOV, int range)
 {
   NewRay::numGridRows = map.getNumGridRows();
   
@@ -294,10 +294,10 @@ void NewRay::performSensingOperation(dummy::Map &map, int posX, int posY, int or
   //std::cout << std::endl << "StartingPhi: " << startingPhi << " EndingPhi: " << endingPhi <<std::endl;
     
   //select the portion of map to be scanned
-  int minI = posX - range;
-  int maxI = posX + range;
-  int minJ = posY - range;
-  int maxJ = posY + range;
+  long minI = posX - range;
+  long maxI = posX + range;
+  long minJ = posY - range;
+  long maxJ = posY + range;
   
   if(minI < 0) minI = 0;
   if(minJ < 0) minJ = 0;
@@ -305,9 +305,9 @@ void NewRay::performSensingOperation(dummy::Map &map, int posX, int posY, int or
   if(maxJ > map.getNumGridCols()) maxJ = map.getNumGridCols();
   
   //scan the cells in the selected portion of the map
-  for(int i = minI; i <= maxI; ++i)
+  for(long i = minI; i <= maxI; ++i)
   {
-    for(int j = minJ; j <=maxJ; ++j)
+    for(long j = minJ; j <=maxJ; ++j)
     {
       
       double distance = sqrt((i - posX)*(i - posX) + (j - posY)*(j - posY));
@@ -347,13 +347,13 @@ void NewRay::performSensingOperation(dummy::Map &map, int posX, int posY, int or
 	    //not needed, but left anyway
 	    if(curX < 0 || curX > map.getNumGridRows() || curY < 0 || curY > map.getNumGridCols()) hit = 1;        
       
-	    if(map.getGridValue((int)curX, (int)curY) == 1) 
+	    if(map.getGridValue((long)curX, (long)curY) == 1) 
 	    {
 	      hit = 1;		//hit set to 1 if an obstacle is found
 	      //std::cout << "HIT! cell: " << j << " " << i << " Hit point: " << curY << " " << curX << std::endl;
 	    }
 
-	    if((int)curX == i && (int)curY == j)	//if the free cell is reached, set its value to 2 and stop the ray
+	    if((long)curX == i && (long)curY == j)	//if the free cell is reached, set its value to 2 and stop the ray
 	    {
 	      map.setGridValue(2, i, j);
 	      //std::cout << "Cell scanned: " << (int)curY << " " << (int)curX << std::endl;
@@ -368,13 +368,13 @@ void NewRay::performSensingOperation(dummy::Map &map, int posX, int posY, int or
 }
 
 //convert the value along the y axis to the cartesian space in order to compute atan2
-int NewRay::convertPoint(int y)
+long NewRay::convertPoint(long y)
 {
   return (NewRay::numGridRows - 1 - y);
 }
 
 
-int NewRay::getInformationGain(const dummy::Map &map, int posX, int posY, int orientation, double FOV, int range)
+int NewRay::getInformationGain(const dummy::Map &map, long posX, long posY, int orientation, double FOV, int range)
 {
   //int raycounter = 0;
   int counter = 0;	//count number of free cells that can be seen
@@ -395,10 +395,10 @@ int NewRay::getInformationGain(const dummy::Map &map, int posX, int posY, int or
   //std::cout << std::endl << "StartingPhi: " << startingPhi << " EndingPhi: " << endingPhi <<std::endl;
     
   //select the portion of map to be scanned
-  int minI = posX - range;
-  int maxI = posX + range;
-  int minJ = posY - range;
-  int maxJ = posY + range;
+  long minI = posX - range;
+  long maxI = posX + range;
+  long minJ = posY - range;
+  long maxJ = posY + range;
   
   if(minI < 0) minI = 0;
   if(minJ < 0) minJ = 0;
@@ -406,9 +406,9 @@ int NewRay::getInformationGain(const dummy::Map &map, int posX, int posY, int or
   if(maxJ > map.getNumGridCols()) maxJ = map.getNumGridCols();
   
   //scan the cells in the selected portion of the map
-  for(int i = minI; i <= maxI; ++i)
+  for(long i = minI; i <= maxI; ++i)
   {
-    for(int j = minJ; j <=maxJ; ++j)
+    for(long j = minJ; j <=maxJ; ++j)
     {
       
       double distance = sqrt((i - posX)*(i - posX) + (j - posY)*(j - posY));
@@ -448,13 +448,13 @@ int NewRay::getInformationGain(const dummy::Map &map, int posX, int posY, int or
 	    //not needed, but left anyway
 	    if(curX < 0 || curX > map.getNumGridRows() || curY < 0 || curY > map.getNumGridCols()) hit = 1;        
       
-	    if(map.getGridValue((int)curX, (int)curY) == 1) 
+	    if(map.getGridValue((long)curX, (long)curY) == 1) 
 	    {
 	      hit = 1;		//hit set to 1 if an obstacle is found
 	      //std::cout << "HIT! cell: " << j << " " << i << " Hit point: " << curY << " " << curX << std::endl;
 	    }
 
-	    if((int)curX == i && (int)curY == j)	//if the free cell is reached, increase counter and stop the ray.
+	    if((long)curX == i && (long)curY == j)	//if the free cell is reached, increase counter and stop the ray.
 	    {
 	      ++counter;
 	      //std::cout << "Cell scanned: " << (int)curY << " " << (int)curX << std::endl;
