+++ ray.h
@@ -14,25 +14,25 @@ public:
   
   Ray();
   //~Ray();
-  void findCandidatePositions(dummy::Map &map, int posX, int posY, int orientation, double FOV, int range);
-  std::vector<std::pair<int, int> > getCandidatePositions();
+  void findCandidatePositions(dummy::Map &map, long posX, long posY, int orientation, double FOV, int range);
+  std::vector<std::pair<long,long> > getCandidatePositions();
   void setGrid(const dummy::Map &map);
-  int getInformationGain(const dummy::Map &map, int posX, int posY, int orientation, double FOV, int range);
-  void performSensingOperation(dummy::Map &map, int posX, int posY, int orientation, double FOV, int range);
-  int getGridValue(int i, int j);
+  int getInformationGain(const dummy::Map &map, long posX, long posY, int orientation, double FOV, int range);
+  void performSensingOperation(dummy::Map &map, long posX, long posY, int orientation, double FOV, int range);
+  int getGridValue(long i, long j);
   void emptyCandidatePositions();
   
 private:
   double mapX, mapY;			//coordinates in the map
-  int posX, posY;		//starting position of the robot
+  long posX, posY;		//starting position of the robot
   int orientation;			//orientation of the robot (0, 90, 180, 270 degrees)
   double FOV;
   int range;			//range of the scanner
-  std::vector<std::pair<int, int> > edgePoints;
-  std::vector<int> grid;
-  void setGridValue(int i, int j, int value);
-  int numGridCols;
-  int numGridRows;
+  std::vector<std::pair<long, long> > edgePoints;
+  std::vector<long> grid;
+  void setGridValue(long i, long j, int value);
+  long numGridCols;
+  long numGridRows;
 };
